# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Spese API
  version: 1.0.0
  description: API JSON same-origin per il ledger personale Spese v2.
servers:
  - url: /
components:
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
...
paths:
...
  /api/v1/accounts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get: { operationId: getAccount, responses: { '200': { description: Conto } } }
    patch: { operationId: updateAccount, parameters: [{ $ref: '#/components/parameters/IfMatch' }], responses: { '200': { description: Conto aggiornato } } }
    delete: { operationId: archiveAccount, parameters: [{ $ref: '#/components/parameters/IfMatch' }], responses: { '200': { description: Conto archiviato } } }
...
  /api/v1/categories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get: { operationId: getCategory, responses: { '200': { description: Categoria } } }
    patch: { operationId: updateCategory, parameters: [{ $ref: '#/components/parameters/IfMatch' }], responses: { '200': { description: Categoria aggiornata } } }
    delete: { operationId: archiveCategory, parameters: [{ $ref: '#/components/parameters/IfMatch' }], responses: { '200': { description: Categoria archiviata } } }
...